refactor(pi-codex-conversion): harden code mode boundaries
